feat(recipe)!: remove invent-from-scratch backend (keep fork + regenerate)
Phase 3a of the finder migration: delete the old "Generate with AI" invent
path — the app now only ever surfaces REAL recipes (search/finder + import).
Fork and regenerate are on different endpoints/methods and stay fully working.

BREAKING: removes POST /v1/recipes/chat and POST /v1/recipes/chat/stream.
Non-updated app builds' "Generate with AI" will 404 there (accepted).

Removed (invent-only):
- internal/service/recipeChat.go — InitGenerateRecipe / FinishGenerateRecipe /
  service StreamGenerateRecipe / finishStreamedRecipe.
- Handlers RecipeHandler.GenerateRecipe + StreamGenerateRecipe (recipe.go); the
  now-unused context/encoding/json/io/time/ai imports dropped.
- Routes POST /recipes/chat + /recipes/chat/stream (router.go).
- The invent tests: service recipe_chat_test.go; and in
  recipe_gating_test.go the TestGenerateRecipe_* / TestStreamGenerateRecipe_*
  cases plus the local closeNotifyRecorder helper and now-unused context/ai
  imports.

Kept working (verified): PUT /recipes/:id/chat (RegenerateRecipe) and
POST /recipes/:id/fork (GenerateRecipeWithFork), the shared
checkAIGenerationLimit / incrementAIGenerationUsage, and the fork/regen service
+ gating tests. Deliberately kept (interface anchors, now production-
unreferenced): TextProvider.GenerateRecipe and StreamingTextProvider.
StreamGenerateRecipe and their impls/mocks — removing them is a large-blast-
radius change with no user benefit.

The shared test helper newGenRecipeService lived in recipe_chat_test.go but is
used by the kept fork/regen service tests, so it was relocated to a new
internal/service/recipe_gen_helpers_test.go (rather than deleted with the file).

Gate: `grep -rn "\.GenerateRecipe(\|\.StreamGenerateRecipe(" --include='*.go'
internal | grep -v _test.go` now shows only the SwitchableTextProvider
passthrough (no invent service/handler caller). go build ./... , go vet ./...
and go test ./... -count=1 all green + offline.
