--- internal/service/recipeChat.go
@@ -1,336 +1,0 @@
-package service
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"time"
-
-	"github.com/windoze95/saltybytes-api/internal/ai"
-	"github.com/windoze95/saltybytes-api/internal/logger"
-	"github.com/windoze95/saltybytes-api/internal/models"
-	"github.com/windoze95/saltybytes-api/internal/util"
-	"go.uber.org/zap"
-)
-
-// InitGenerateRecipe initializes a new recipe with chat.
-func (s *RecipeService) InitGenerateRecipe(user *models.User, userPrompt string, genImage bool) (*RecipeResponse, error) {
-	if user.Personalization.ID == 0 {
-		logger.Get().Warn("user personalization is nil", zap.Uint("user_id", user.ID))
-		return nil, errors.New("user's Personalization is nil")
-	}
-
-	// Populate initial fields of the Recipe struct
-	recipe := &models.Recipe{
-		CreatedBy:          user,
-		PersonalizationUID: user.Personalization.UID,
-		Status:             "generating",
-	}
-
-	// Create a Recipe with the basic Recipe details
-	if err := s.Repo.CreateRecipe(recipe); err != nil {
-		return nil, fmt.Errorf("failed to save recipe record: %w", err)
-	}
-
-	recipeResponse := s.ToRecipeResponse(recipe)
-
-	go s.FinishGenerateRecipe(recipe, user, userPrompt, genImage)
-
-	// The recipe now has an ID generated by the database
-	return recipeResponse, nil
-}
-
-// FinishGenerateRecipe finishes generating a recipe with chat.
-func (s *RecipeService) FinishGenerateRecipe(recipe *models.Recipe, user *models.User, userPrompt string, genImage bool) {
-	defer util.RecoverPanic("finish generate recipe")
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	defer cancel()
-
-	// Both channels are buffered so the worker goroutines' single send never
-	// blocks (and leaks the goroutine) when the parent returns on ctx.Done().
-	recipeErrChan := make(chan error, 1)
-	imageErrChan := make(chan error, 1)
-
-	req := ai.RecipeRequest{
-		UserPrompt:     userPrompt,
-		UnitSystem:     user.Personalization.UnitSystemText(),
-		Requirements:   user.Personalization.Requirements,
-		CookingContext: user.Personalization.CookingContextPrompt(),
-	}
-
-	// Goroutine to handle recipe generation
-	go func(ctx context.Context, recipeErrChan chan<- error, imageErrChan chan<- error) {
-		defer util.RecoverPanic("generate recipe worker")
-
-		result, err := s.TextProvider.GenerateRecipe(ctx, req)
-		if err != nil {
-			recipeErrChan <- err
-			return
-		}
-
-		if result.UnitSystem == "" {
-			result.UnitSystem = user.Personalization.UnitSystem
-		}
-		recipeDef := recipeResultToRecipeDef(result)
-		recipe.RecipeDef = recipeDef
-		recipe.PromptVersion = result.PromptVersion
-
-		if err := validateRecipeCoreFields(recipe); err != nil {
-			recipeErrChan <- err
-			return
-		}
-
-		// Goroutine to handle image generation and upload
-		go func(ctx context.Context, imageErrChan chan<- error) {
-			defer util.RecoverPanic("generate recipe image worker")
-
-			if genImage && result.ImagePrompt != "" {
-				imageBytes, imgErr := s.ImageProvider.GenerateImage(ctx, result.ImagePrompt)
-				if imgErr != nil {
-					imageErrChan <- imgErr
-					return
-				}
-
-				imageURL, uploadErr := uploadRecipeImage(ctx, recipe.ID, recipe.ImageURL, imageBytes, s.Cfg)
-				if uploadErr != nil {
-					imageErrChan <- uploadErr
-					return
-				}
-
-				if dbErr := s.Repo.UpdateRecipeImageURL(recipe.ID, imageURL); dbErr != nil {
-					imageErrChan <- dbErr
-					return
-				}
-			}
-
-			imageErrChan <- nil
-		}(ctx, imageErrChan)
-
-		if err := s.Repo.UpdateRecipeDef(recipe); err != nil {
-			recipeErrChan <- err
-			return
-		}
-
-		if err := s.AssociateTagsWithRecipe(recipe, result.Hashtags); err != nil {
-			logger.Get().Error("failed to associate tags with recipe", zap.Uint("recipe_id", recipe.ID), zap.Error(err))
-		}
-
-		// Create a tree for the new recipe with the initial result as the root node
-		rootNode := &models.RecipeNode{
-			Prompt:      userPrompt,
-			Response:    &recipeDef,
-			Summary:     result.Summary,
-			Type:        models.RecipeTypeChat,
-			BranchName:  "original",
-			CreatedByID: recipe.CreatedByID,
-			IsActive:    true,
-		}
-		if _, err := s.Repo.CreateRecipeTree(recipe.ID, rootNode); err != nil {
-			logger.Get().Error("failed to create recipe tree", zap.Uint("recipe_id", recipe.ID), zap.Error(err))
-			// Non-fatal: the recipe was created successfully, tree is supplementary
-		}
-
-		s.generateAndStoreEmbedding(ctx, recipe.ID, &recipeDef)
-
-		s.Repo.UpdateRecipeStatus(recipe.ID, "ready")
-
-		recipeErrChan <- nil
-	}(ctx, recipeErrChan, imageErrChan)
-
-	// Wait for the recipe generation goroutine to finish or timeout
-	select {
-	case err := <-recipeErrChan:
-		if err != nil {
-			recipeID := recipe.ID
-			logger.Get().Error("failed to finish recipe generation", zap.Uint("recipe_id", recipeID), zap.Error(err))
-			s.Repo.UpdateRecipeStatus(recipeID, "failed")
-			e := s.DeleteRecipe(context.Background(), recipeID)
-			if e != nil {
-				logger.Get().Error("failed to delete recipe after generation error", zap.Uint("recipe_id", recipeID), zap.Error(e))
-				return
-			}
-			logger.Get().Info("recipe deleted after generation error", zap.Uint("recipe_id", recipeID))
-			return
-		}
-	case <-ctx.Done():
-		err := errors.New("incomplete recipe generation: timed out after 5 minutes")
-		recipeID := recipe.ID
-		logger.Get().Error("recipe generation timed out", zap.Uint("recipe_id", recipeID), zap.Error(err))
-		s.Repo.UpdateRecipeStatus(recipeID, "failed")
-		e := s.DeleteRecipe(context.Background(), recipeID)
-		if e != nil {
-			logger.Get().Error("failed to delete recipe after timeout", zap.Uint("recipe_id", recipeID), zap.Error(e))
-			return
-		}
-		logger.Get().Info("recipe deleted after timeout", zap.Uint("recipe_id", recipeID))
-		return
-	}
-
-	if !genImage {
-		return
-	}
-
-	// Wait for the image generation goroutine to finish or timeout
-	select {
-	case err := <-imageErrChan:
-		if err != nil {
-			logger.Get().Error("failed to generate recipe image", zap.Uint("recipe_id", recipe.ID), zap.Error(err))
-			return
-		}
-	case <-ctx.Done():
-		logger.Get().Error("recipe image generation timed out", zap.Uint("recipe_id", recipe.ID))
-		return
-	}
-}
-
-// StreamGenerateRecipe creates and generates a recipe using streaming, emitting
-// progress events to the provided channel. The caller must close the channel
-// after this method returns.
-func (s *RecipeService) StreamGenerateRecipe(ctx context.Context, user *models.User, userPrompt string, genImage bool, events chan<- ai.StreamEvent) {
-	log := logger.Get().With(zap.Uint("user_id", user.ID))
-
-	if user.Personalization.ID == 0 {
-		log.Warn("user personalization is nil")
-		ai.TrySendEvent(ctx, events, ai.StreamEvent{Type: ai.StreamEventError, Error: "user personalization is nil", ErrorKind: "content_quality"})
-		return
-	}
-
-	// Create placeholder recipe
-	recipe := &models.Recipe{
-		CreatedBy:          user,
-		PersonalizationUID: user.Personalization.UID,
-		Status:             "generating",
-	}
-	if err := s.Repo.CreateRecipe(recipe); err != nil {
-		log.Error("failed to save recipe record", zap.Error(err))
-		ai.TrySendEvent(ctx, events, ai.StreamEvent{Type: ai.StreamEventError, Error: "failed to create recipe", ErrorKind: "unknown"})
-		return
-	}
-
-	ai.TrySendEvent(ctx, events, ai.StreamEvent{Type: ai.StreamEventStarted, RecipeID: recipe.ID})
-
-	// Type-assert to get streaming capability
-	streamProvider, ok := s.TextProvider.(ai.StreamingTextProvider)
-	if !ok {
-		// Fall back to non-streaming generation
-		log.Info("provider does not support streaming, falling back to sync generation")
-		result, err := s.TextProvider.GenerateRecipe(ctx, ai.RecipeRequest{
-			UserPrompt:     userPrompt,
-			UnitSystem:     user.Personalization.UnitSystemText(),
-			Requirements:   user.Personalization.Requirements,
-			CookingContext: user.Personalization.CookingContextPrompt(),
-		})
-		if err != nil {
-			log.Error("sync recipe generation failed", zap.Uint("recipe_id", recipe.ID), zap.Error(err))
-			s.Repo.UpdateRecipeStatus(recipe.ID, "failed")
-			s.DeleteRecipe(context.Background(), recipe.ID)
-			ai.TrySendEvent(ctx, events, ai.StreamEvent{Type: ai.StreamEventError, RecipeID: recipe.ID, Error: err.Error(), ErrorKind: "unknown"})
-			return
-		}
-		s.finishStreamedRecipe(ctx, recipe, user, userPrompt, result, genImage, events)
-		return
-	}
-
-	req := ai.RecipeRequest{
-		UserPrompt:     userPrompt,
-		UnitSystem:     user.Personalization.UnitSystemText(),
-		Requirements:   user.Personalization.Requirements,
-		CookingContext: user.Personalization.CookingContextPrompt(),
-	}
-
-	result, err := streamProvider.StreamGenerateRecipe(ctx, req, events)
-	if err != nil {
-		log.Error("streaming recipe generation failed", zap.Uint("recipe_id", recipe.ID), zap.Error(err))
-		s.Repo.UpdateRecipeStatus(recipe.ID, "failed")
-		s.DeleteRecipe(context.Background(), recipe.ID)
-		// Error event already emitted by StreamGenerateRecipe
-		return
-	}
-
-	s.finishStreamedRecipe(ctx, recipe, user, userPrompt, result, genImage, events)
-}
-
-// finishStreamedRecipe persists the generated recipe and emits the complete event.
-func (s *RecipeService) finishStreamedRecipe(ctx context.Context, recipe *models.Recipe, user *models.User, userPrompt string, result *ai.RecipeResult, genImage bool, events chan<- ai.StreamEvent) {
-	log := logger.Get().With(zap.Uint("recipe_id", recipe.ID))
-
-	if result.UnitSystem == "" {
-		result.UnitSystem = user.Personalization.UnitSystem
-	}
-	recipeDef := recipeResultToRecipeDef(result)
-	recipe.RecipeDef = recipeDef
-	recipe.PromptVersion = result.PromptVersion
-
-	if err := validateRecipeCoreFields(recipe); err != nil {
-		log.Error("recipe validation failed", zap.Error(err))
-		s.Repo.UpdateRecipeStatus(recipe.ID, "failed")
-		s.DeleteRecipe(context.Background(), recipe.ID)
-		ai.TrySendEvent(ctx, events, ai.StreamEvent{Type: ai.StreamEventError, RecipeID: recipe.ID, Error: err.Error(), ErrorKind: "content_quality"})
-		return
-	}
-
-	if err := s.Repo.UpdateRecipeDef(recipe); err != nil {
-		log.Error("failed to persist recipe", zap.Error(err))
-		s.Repo.UpdateRecipeStatus(recipe.ID, "failed")
-		ai.TrySendEvent(ctx, events, ai.StreamEvent{Type: ai.StreamEventError, RecipeID: recipe.ID, Error: "failed to save recipe", ErrorKind: "unknown"})
-		return
-	}
-
-	if err := s.AssociateTagsWithRecipe(recipe, result.Hashtags); err != nil {
-		log.Error("failed to associate tags", zap.Error(err))
-	}
-
-	rootNode := &models.RecipeNode{
-		Prompt:      userPrompt,
-		Response:    &recipeDef,
-		Summary:     result.Summary,
-		Type:        models.RecipeTypeChat,
-		BranchName:  "original",
-		CreatedByID: recipe.CreatedByID,
-		IsActive:    true,
-	}
-	if _, err := s.Repo.CreateRecipeTree(recipe.ID, rootNode); err != nil {
-		log.Error("failed to create recipe tree", zap.Error(err))
-	}
-
-	s.generateAndStoreEmbedding(ctx, recipe.ID, &recipeDef)
-	s.Repo.UpdateRecipeStatus(recipe.ID, "ready")
-
-	// Image generation runs in background — SSE closes after complete
-	if genImage && result.ImagePrompt != "" {
-		go func() {
-			defer util.RecoverPanic("streamed recipe background image")
-
-			imgCtx, cancel := context.WithTimeout(context.Background(), 2*time.Minute)
-			defer cancel()
-			imageBytes, err := s.ImageProvider.GenerateImage(imgCtx, result.ImagePrompt)
-			if err != nil {
-				log.Error("background image generation failed", zap.Error(err))
-				return
-			}
-			imageURL, err := uploadRecipeImage(imgCtx, recipe.ID, recipe.ImageURL, imageBytes, s.Cfg)
-			if err != nil {
-				log.Error("background image upload failed", zap.Error(err))
-				return
-			}
-			s.Repo.UpdateRecipeImageURL(recipe.ID, imageURL)
-		}()
-	}
-
-	recipeResponse := s.ToRecipeResponse(recipe)
-	ai.TrySendEvent(ctx, events, ai.StreamEvent{Type: ai.StreamEventComplete, RecipeID: recipe.ID, Result: &ai.RecipeResult{
-		Title:             recipeResponse.Title,
-		Ingredients:       result.Ingredients,
-		Instructions:      result.Instructions,
-		CookTime:          result.CookTime,
-		ImagePrompt:       result.ImagePrompt,
-		Hashtags:          result.Hashtags,
-		LinkedSuggestions: result.LinkedSuggestions,
-		Summary:           result.Summary,
-		Portions:          result.Portions,
-		PortionSize:       result.PortionSize,
-		UnitSystem:        result.UnitSystem,
-		PromptVersion:     result.PromptVersion,
-	}})
-}
